Move endpoint convention builder
