+++ src/Components/Server/src/Builder/ComponentEndpointConventionBuilder.cs
@@ -2,10 +2,9 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.SignalR;
 
-namespace Microsoft.AspNetCore.Components.Server
+namespace Microsoft.AspNetCore.Builder
 {
     /// <summary>
     /// Builds conventions that will be used for customization of ComponentHub <see cref="EndpointBuilder"/> instances.
